+++ src/brokercollect/broker_collection.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/newrelic/infra-integrations-sdk/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/integration"
+	"github.com/newrelic/infra-integrations-sdk/jmx"
 	"github.com/newrelic/infra-integrations-sdk/log"
 	"github.com/newrelic/nri-kafka/src/args"
 	"github.com/newrelic/nri-kafka/src/jmxwrapper"
@@ -116,12 +117,13 @@ func createBroker(brokerID int, zkConn zookeeper.Connection, i *integration.Inte
 	// Collect broker connection information from ZooKeeper
 	_, host, jmxPort, kafkaPort, err := zookeeper.GetBrokerConnectionInfo(brokerID, zkConn)
 	if err != nil {
-		log.Error("Unable to get broker JMX information for broker id %s: %s", host, err)
+		log.Error("Unable to get broker JMX information for broker id %s: %s", brokerID, err)
 		return nil, err
 	}
 
 	// Create broker entity
-	brokerEntity, err := i.Entity(host, "broker")
+	clusterIDAttr := integration.NewIDAttribute("clusterName", args.GlobalArgs.ClusterName)
+	brokerEntity, err := i.Entity(host, "ka-broker", clusterIDAttr)
 	if err != nil {
 		log.Error("Unable to create entity for broker ID %d: %s", brokerID, err)
 		return nil, err
@@ -173,7 +175,13 @@ func collectBrokerMetrics(b *broker, collectedTopics []string) error {
 	jmxwrapper.JMXLock.Lock()
 
 	// Open JMX connection
-	if err := jmxwrapper.JMXOpen(b.Host, strconv.Itoa(b.JMXPort), args.GlobalArgs.DefaultJMXUser, args.GlobalArgs.DefaultJMXPassword); err != nil {
+	options := make([]jmx.Option, 0)
+	if args.GlobalArgs.KeyStore != "" && args.GlobalArgs.KeyStorePassword != "" && args.GlobalArgs.TrustStore != "" && args.GlobalArgs.TrustStorePassword != "" {
+		ssl := jmx.WithSSL(args.GlobalArgs.KeyStore, args.GlobalArgs.KeyStorePassword, args.GlobalArgs.TrustStore, args.GlobalArgs.TrustStorePassword)
+		options = append(options, ssl)
+	}
+
+	if err := jmxwrapper.JMXOpen(b.Host, strconv.Itoa(b.JMXPort), args.GlobalArgs.DefaultJMXUser, args.GlobalArgs.DefaultJMXPassword, options...); err != nil {
 		log.Error("Unable to make JMX connection for Broker '%s': %s", b.Host, err.Error())
 		jmxwrapper.JMXClose() // Close needs to be called even on a failed open to clear out any set variables
 		jmxwrapper.JMXLock.Unlock()