+++ src/prodconcollect/producer_consumer_collection_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/newrelic/infra-integrations-sdk/integration"
+	"github.com/newrelic/infra-integrations-sdk/jmx"
 	"github.com/newrelic/nri-kafka/src/args"
 	"github.com/newrelic/nri-kafka/src/jmxwrapper"
 	"github.com/newrelic/nri-kafka/src/testutils"
@@ -85,7 +86,9 @@ func TestConsumerWorker(t *testing.T) {
 
 func TestConsumerWorker_JmxOpenFuncErr(t *testing.T) {
 	testutils.SetupJmxTesting()
-	jmxwrapper.JMXOpen = func(hostname, port, username, password string) error { return errors.New("test") }
+	jmxwrapper.JMXOpen = func(hostname, port, username, password string, options ...jmx.Option) error {
+		return errors.New("test")
+	}
 	consumerChan := make(chan *args.JMXHost, 10)
 	var wg sync.WaitGroup
 	i, err := integration.New("kafka", "1.0.0")
@@ -135,7 +138,9 @@ func TestProducerWorker(t *testing.T) {
 
 func TestProducerWorker_JmxOpenFuncErr(t *testing.T) {
 	testutils.SetupJmxTesting()
-	jmxwrapper.JMXOpen = func(hostname, port, username, password string) error { return errors.New("test") }
+	jmxwrapper.JMXOpen = func(hostname, port, username, password string, options ...jmx.Option) error {
+		return errors.New("test")
+	}
 	producerChan := make(chan *args.JMXHost, 10)
 	var wg sync.WaitGroup
 	i, err := integration.New("kafka", "1.0.0")