+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.2.1 - 2019-07-15
+### Changed
+- Added documentation in sample config file for regex list mode
+
+## 2.2.0 - 2019-07-11
+### Added
+- Regex list mode for Kafka
+
+## 2.1.0 - 2019-06-25
+### Added
+- SSL support for JMX connections
+
+## 2.0.0 - 2019-04-22
+### Changed
+- Changed the entity namespaces to be kafka-scoped
+- ClusterName is a required argument to better enforce uniquene
+
 ## 1.1.1 - 2019-02-04
 ### Changed
 - Updated Definition file protocol version to 2